backend_services/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
click==8.1.7
colorama==0.4.6
fastapi==0.103.1
h11==0.14.0
idna==3.4
psycopg2==2.9.7
pydantic==2.3.0
pydantic_core==2.6.3
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.23.2
httpx==0.24.1
websockets==11.0.3
requests==2.26.0
python-dotenv==1.0.0
